refactor: change component folder structure and add styles
