+++ src/App.js
@@ -1,38 +1,48 @@
-import React from 'react';
+import React  from 'react';
 import axios from 'axios';
 
-import Search from './Components/Search';
-import City from './Components/City';
+import Header from './Components/Header/Header';
+import Search from './Components/Search/Search';
+import City from './Components/City/City';
+
+import './App.css'
+
 class App extends React.Component {
   constructor(props) {
     super(props);
     this.state={
       haveSearched: false,
       citySearched: '',
+      locationData: {},
     };
   }
+
+  
 
   showSearch = () => {
     this.setState({haveSearched: false});
   }
 
   handleSearch = async(citySearched) => {
-    // TODO: Handle searches from multiple regions
-    let response = await axios.get(`https://us1.locationiq.com/v1/search.php?key=${process.env.REACT_APP_LOCATION_IQ_API_KEY}&q=${citySearched}&format=json&limit=1`);
-    console.log(response);
-    this.setState({
-      haveSearched: true,
-      citySearched: citySearched,
-      locationData: response.data[0],
-    });
+    if(!citySearched) {
+      console.warn('No City Selected');
+    } else {
+      // TODO: Handle searches from multiple regions
+      let response = await axios.get(`https://us1.locationiq.com/v1/search.php?key=${process.env.REACT_APP_LOCATION_IQ_API_KEY}&q=${citySearched}&format=json&limit=1`);
+      this.setState({
+        haveSearched: true,
+        citySearched: citySearched,
+        locationData: response.data[0],
+      });
+    }
   }
 
   render() {
     return (
       <>
-        <h1>City Explorer</h1>
-        {this.state.haveWeSearchedYet ? 
-          <City handleShowSearch={this.handleShowSearch} cityData={this.state.locationData} /> : 
+        <Header />
+        {this.state.haveSearched ? 
+          <City handleShowSearch={this.showSearch} cityData={this.state.locationData} /> : 
           <Search handleSearch={this.handleSearch} />}
       </>
     );
