+++ src/Components/City/City.css
@@ -1,0 +1,4 @@
+Button {
+  width: 100vw;
+  text-align: center;
+}
