+++ src/Components/Header/Header.css
@@ -1,0 +1,5 @@
+header {
+  text-align: center;
+  background-color: rgba(187, 214, 177, 0.472);
+  color :#304634
+}
