+++ src/Components/Header/Header.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import './Header.css';
+
+class Header extends React.Component {
+  render() {
+    return (
+      <header>
+        <h1>City Explorer</h1>
+      </header>
+    );
+  }
+}
+
+export default Header;
