+++ src/Components/Search/Search.css
@@ -1,0 +1,8 @@
+Form label, Form Button {
+  width: 100vw;
+  text-align: center;
+}
+
+Form label {
+  font-size: 1.5rem;
+}
