+++ src/Components/Search/Search.js
@@ -1,0 +1,34 @@
+import React from 'react';
+
+import 'bootstrap/dist/css/bootstrap.min.css';
+import { Form, Button } from 'react-bootstrap';
+
+import './Search.css';
+
+class Search extends React.Component {
+  constructor(props) {
+    super(props);
+    this.textInput = React.createRef();
+  }
+
+  handleSubmit = (event) => {
+    event.preventDefault();
+    this.props.handleSearch(this.textInput.current.value);
+  }
+
+  render() {
+    return (
+      <Form onSubmit={this.handleSubmit}>
+        <Form.Group>
+          <Form.Label>Enter City to Search For</Form.Label>
+          <Form.Control placeholder="Enter city to search..." size="lg" type="text" ref={this.textInput} />
+        </Form.Group>
+        <Button variant="primary" type="submit">
+          Explore!
+        </Button>
+      </Form>
+    );
+  }
+}
+
+export default Search
